# speedup server tests by reducing checks from 30 seconds to 1 second