Before
```
112.94s call     tests/server/test_slurm_server.py::test_gcd_server
 85.01s call     tests/server/test_server.py::test_gcd_server
 84.55s call     tests/server/test_gcd_server_auth.py::test_gcd_server_authenticated
```

After
```
85.50s call     tests/server/test_slurm_server.py::test_gcd_server
62.80s call     tests/server/test_server.py::test_gcd_server
51.81s call     tests/server/test_gcd_server_auth.py::test_gcd_server_authenticated
```